fix(files): refactoring
Removed sponsors and teams arrays from .svelte files and moved them to
$lib/arrays for better accessibility
