+++ src/lib/arrays/sponsors.js
@@ -1,0 +1,10 @@
+// Sponsors data for the new section
+let sponsors = [
+  {
+    name: "Hack Club",
+    logo: "https://assets.hackclub.com/icon-rounded.png",
+    url: "https://hackclub.com",
+  },
+];
+
+export default sponsors
