+++ src/lib/arrays/team.js
@@ -1,0 +1,19 @@
+let team = [
+  {
+    Name: "Drew",
+    Avatar: "https://tbilisi.hackclub.com/drew/avatar.png",
+    Url: "https://tbilisi.hackclub.com/drew/",
+  },
+  {
+    Name: "Boris Mirakov",
+    Avatar: "https://tbilisi.hackclub.com/bormir/avatar.png",
+    Url: "https://tbilisi.hackclub.com/bormir/",
+  },
+  {
+    Name: "Zlata Chernova",
+    Avatar: "https://tbilisi.hackclub.com/zlataponchikk/avatar.png",
+    Url: "https://tbilisi.hackclub.com/zlataponchikk",
+  },
+];
+
+export default team
