+++ src/routes/+page.svelte
@@ -1,17 +1,5 @@
 <script lang="js">
-    // Sponsors data for the new section
-    let sponsors = [
-        {
-            name: "Hack Club",
-            logo: "https://assets.hackclub.com/icon-rounded.png",
-            url: "https://hackclub.com",
-        },
-        {
-            name: "Drew Smirnoff",
-            logo: "https://tbilisi.hackclub.com/drew/avatar.png",
-            url: "https://tbilisi.hackclub.com/drew/",
-        },
-    ];
+    import sponsors from "$lib/arrays/sponsors.js";
 </script>
 
 <svelte:head>
