+++ src/routes/about/+page.svelte
@@ -1,21 +1,5 @@
 <script lang="js">
-    let team = [
-        {
-            Name: "Drew",
-            Avatar: "https://tbilisi.hackclub.com/drew/avatar.png",
-            Url: "https://tbilisi.hackclub.com/drew/",
-        },
-        {
-            Name: "Boris Mirakov",
-            Avatar: "https://tbilisi.hackclub.com/bormir/avatar.png",
-            Url: "https://tbilisi.hackclub.com/bormir/",
-        },
-        {
-            Name: "Zlata Chernova",
-            Avatar: "https://tbilisi.hackclub.com/zlataponchikk/avatar.png",
-            Url: "https://tbilisi.hackclub.com/zlataponchikk",
-        },
-    ];
+    import team from "$lib/arrays/team.js";
 
     // Fallback image handler
     function handleImgError(event, name) {
